Move Result to the remorph package
