+++ core/src/main/scala/com/databricks/labs/remorph/discovery/Anonymizer.scala
@@ -2,8 +2,8 @@ package com.databricks.labs.remorph.discovery
 
 import com.databricks.labs.remorph.parsers.PlanParser
 import com.databricks.labs.remorph.intermediate._
-import com.databricks.labs.remorph.transpilers.WorkflowStage.PARSE
-import com.databricks.labs.remorph.transpilers.{Result, SourceCode}
+import com.databricks.labs.remorph.{Result, WorkflowStage}
+import com.databricks.labs.remorph.transpilers.SourceCode
 import com.typesafe.scalalogging.LazyLogging
 import upickle.default._
 
@@ -72,7 +72,7 @@ class Anonymizer(parser: PlanParser[_]) extends LazyLogging {
 
   private[discovery] def fingerprint(query: ExecutedQuery): Fingerprint = {
     parser.parse(SourceCode(query.source)).flatMap(parser.visit) match {
-      case Result.Failure(PARSE, errorJson) =>
+      case Result.Failure(WorkflowStage.PARSE, errorJson) =>
         logger.warn(s"Failed to parse query: ${query.source} $errorJson")
         Fingerprint(
           query.id,
