+++ core/src/main/scala/com/databricks/labs/remorph/parsers/PlanParser.scala
@@ -1,6 +1,7 @@
 package com.databricks.labs.remorph.parsers
 
-import com.databricks.labs.remorph.transpilers.{Result, SourceCode, WorkflowStage}
+import com.databricks.labs.remorph.{Result, WorkflowStage}
+import com.databricks.labs.remorph.transpilers.SourceCode
 import com.databricks.labs.remorph.{intermediate => ir}
 import org.antlr.v4.runtime._
 import org.json4s.jackson.Serialization
